requirements.txt
tqdm
pydantic
PyPDF2
pymupdf
python-dotenv

langchain
langchain-openai

azure-ai-documentintelligence
azure-core
